[SPARK-53001][FOLLOW-UP] Integrate RocksDB Memory Usage with the Unified Memory Manager
